+++ 2nd_Team2/2nd_Team2/Line.cpp
@@ -29,6 +29,6 @@ void CLine::Render(HDC hDC)
 	float fScorll_X = CScrollMgr::Get_Scroll()->Get_ScrollX();
 	float fScorll_Y = CScrollMgr::Get_Scroll()->Get_ScrollY();
 
-	MoveToEx(hDC, m_tInfo.tLPoint.fX + fScorll_X, m_tInfo.tLPoint.fY + fScorll_Y, nullptr);
-	LineTo(hDC, m_tInfo.tRPoint.fX + fScorll_X, m_tInfo.tRPoint.fY + fScorll_Y);
+	MoveToEx(hDC, static_cast<int>(m_tInfo.tLPoint.fX + fScorll_X), static_cast<int>(m_tInfo.tLPoint.fY + fScorll_Y), nullptr);
+	LineTo(hDC, static_cast<int>(m_tInfo.tRPoint.fX + fScorll_X), static_cast<int>(m_tInfo.tRPoint.fY + fScorll_Y));
 }
